requirements.txt
openai==0.27.0
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-dotenv==1.0.0
pydantic==2.0.3